Some changes

+++ schema.py
@@ -1,6 +1,5 @@
 from pydantic import BaseModel
 from datetime import date
-from decimal import Decimal
 
 
 class SearchDataResponse(BaseModel):
@@ -8,5 +7,5 @@ class SearchDataResponse(BaseModel):
     cadastral_number: str
     date: date
     normative_act: str
-    cadastral_value: Decimal
+    cadastral_value: float
     note: str
